requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
shapely
pydantic
supabase
python-dotenv
